machine-learning-client/requirement.txt
flask
flask-cors
fer
pymongo
opencv-python
Pillow
numpy==1.26.4
moviepy
tensorflow>=1.7